Add common relationship validators
